Fix for Size and Symbols not found issues.
This is to fix the issued mentioned in the following issue as the first and third problem:

https://github.com/prsolucoes/mobile-pdfium/issues/1

With this change I am able to achieve a 38 Mb universal .a file rather than a gigabyte file.
Linker errors; which were thrown during build of application using this library; are also fixed.

Other two problems mentioned in the issue are not related to the project I am working on. I will not be able to work on them.